pyproject.toml
[project]
name = "autoresearch-team"
version = "0.1.0"
description = "Collective intelligence layer for autonomous ML research"
requires-python = ">=3.11"
dependencies = [
    "lightning-sdk",
    "pyyaml",
    "click>=8.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["infra"]

[project.scripts]
art = "infra.lightning.cli:cli"

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-timeout>=2.2",
]

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests requiring live Lightning AI Studios (slow, requires auth)",
]
testpaths = ["tests"]
timeout = 600
